devtools: reconfigure symbol-check for Dogecoin 1.14 targets
Reconfigures the Bitcoin 24.0.1 symbol-check.py script to honor
the maximum versions of dynamic symbols, the allowed system
dependencies and allowed symbol exports. This is important to
maintain when doing minor releases, because changes in these
would potentially lock people out of security updates.

This adds specification of the linker-loader name for i686
binaries because Bitcoin Core no longer supports that
architecture.

The spec was taken from:
https://sourceware.org/glibc/wiki/ABIList?action=recall&rev=16

Please note that:
- aarch64 binaries have had a glibc 2.17 requirement since the
  first release with 1.14.0, and therefore have a higher glibc
  target than all other linux binaries.
- All other values have been taken from the Dogecoin Core v1.14.6
  tag, commit 3a29ba6d4.
- Additional win32 and win64 needed libraries have been reverse
  engineered from 1.14.6 release binaries.
- Windows minimum version checks have been disabled, as these
  need to be set on the release binaries before we check for it.