build: clean up security and symbol checks from makefile
Fixes calls to make check-security and make check-symbols
to have better integration with the CI and Gitian

The condition in the check-symbols target requiring the configure
flag --enable-glibc-back-compat is removed because that is
exclusive to Linux. By removing it, we enable the check for all
binaries, including those built for Windows and macOs.

Removes READELF and CPPFILT variables throughout because these
are not used anywhere after this change