Allow PaymentServer a parametric server name
While the default server name is almost always what we want, given that
it is unique to main or testnet and the default data directory,
defaulting to this value during tests is not always what we want,
especially if there's production data or the path is not available
(think of a CI system).

Switch the test to provide a fixed server name at the point of
constructing the PaymentServer object.

Addresses GH #3285.